// test/enums/common.ts
/* eslint-disable @typescript-eslint/no-unused-vars */
// Note: Disabled the @typescript-eslint/no-unused-vars because this file will only contain the enums
// and due to an issue with eslint it is unable to detect the enum usage

//---------------------------------- Party and Account enums ----------------------------------
export enum AccountPartyType {
        NATURAL_MULTIPLE_JOINT = "Natural/Multiple/Joint",
        NON_NATURAL = "Non-natural"
}

export enum AttributeName {
        accept = "accept",
        alt = "alt",
        arialabel = "aria-label",
        autocomplete = "autocomplete",
        checked = "checked",
        class = "class",
        disabled = "disabled",
        href = "href",
        id = "id",
        max = "max",
        maxlength = "maxlength",
        method = "method",
        min = "min",
        multiple = "multiple",
        name = "name",
        placeholder = "placeholder",
        readonly = "readonly",
        required = "required",
        selected = "selected",
        src = "src",
        style = "style",
        type = "type",
        value = "value",
        text = "text",
        title = "title",
}